+++ src/main/java/com/frejdh/util/common/functional/ThrowingFunction.java
@@ -2,6 +2,8 @@ package com.frejdh.util.common.functional;
 
 import java.util.function.Function;
 
+import static com.frejdh.util.common.toolbox.CommonUtils.sneakyThrow;
+
 @FunctionalInterface
 public interface ThrowingFunction<T, R> extends Function<T, R> {
 	@Override
@@ -9,7 +11,8 @@ public interface ThrowingFunction<T, R> extends Function<T, R> {
 		try {
 			return applyThrows(parameter);
 		} catch (final Exception e) {
-			throw new RuntimeException(e);
+			sneakyThrow(e);
+			throw new RuntimeException(e); // Never reached, but required for compilation nevertheless
 		}
 	}