+++ src/main/java/com/frejdh/util/common/invocations/EqualsValueCondition.java
@@ -2,10 +2,10 @@ package com.frejdh.util.common.invocations;
 
 import java.util.List;
 
-public class EqualsCondition<T> extends Condition<T> {
+public class EqualsValueCondition<T> extends AbstractCondition<T> {
 	protected List<T> equalsToValue;
 
-	public EqualsCondition(ThrowableUtils<T> parent, List<T> equalsToValue) {
+	public EqualsValueCondition(Conditionals<T> parent, List<T> equalsToValue) {
 		this.parent = parent;
 		this.equalsToValue = equalsToValue;
 	}