+++ src/main/java/com/frejdh/util/common/invocations/ThrowableCondition.java
@@ -2,10 +2,10 @@ package com.frejdh.util.common.invocations;
 
 import java.util.List;
 
-public class ThrowableCondition<T> extends Condition<T> {
+public class ThrowableCondition<T> extends AbstractCondition<T> {
 	protected List<Class<? extends Throwable>> throwableClasses;
 
-	public ThrowableCondition(ThrowableUtils<T> parent, List<Class<? extends Throwable>> throwableClasses) {
+	public ThrowableCondition(Conditionals<T> parent, List<Class<? extends Throwable>> throwableClasses) {
 		this.parent = parent;
 		this.throwableClasses = throwableClasses;
 	}